add recycler views for hourly and daily forecast on main screen (mvp)

+++ app/src/main/java/com/example/weatherapp/network/model/multiple_forecast_data/CurrentDto.kt
@@ -23,7 +23,7 @@ data class CurrentDto(
     @SerializedName("temp")
     val temp: Double?,
     @SerializedName("uvi")
-    val uvi: Int?,
+    val uvi: Double?,
     @SerializedName("visibility")
     val visibility: Int?,
     @SerializedName("weather")
